requirements/local.txt
-r base.txt

django-debug-toolbar==4.2.0
django-extensions==3.2.3
docutils==0.20.1
pytest==7.4.4
pytest-django==4.7.0
